LayoutTests/fast/scrolling/ios/scroll-to-beginning-and-end-of-document-expected.txt

Test key commands to scroll to the end of the document and then to the beginning of the document. To run this test by hand, reload the page then press Command + Down Arrow and then press Command + Up Arrow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.scrollY is INITIAL_Y_OFFSET
PASS window.scrollY is >= INITIAL_Y_OFFSET + 1
PASS window.scrollY is 0
PASS successfullyParsed is true

TEST COMPLETE
